Make each channel contain a list of autochans instead of having one global list.
This addresses Vadim's idea on review board. Since ABI compatibility is already being
broken with the ast_channel_ao2 branch, we can afford to add the autochan list to the
channel structure directly.

I couldn't test this since app_dahdiscan is not compiling. I'll get that fixed and commit
that, too.