Fixed autologoff in agents.conf not working when agent logs in via AgentLogin app
An agent logs in by calling an extension that calls the AgentLogin app.  In agents.conf ackcall=always is set, so when they get a call they have the choice to either acknowledge it or ignore it.  autologoff=10 is set as well, so if the agent ignores the call over 10sec one may assume that the agent should be logged out (and in this case hungup on as well), but this was not happening.

(closes issue #14091)
Reported by: evandro
Patches:
      autologoff.diff uploaded by dvossel (license 671)

Review: http://reviewboard.digium.com/r/225/