Add attendee (read) support for iCalendar and CalDAV
Of course, MS in their infinite wisdom, don't make the attendee list
available over WebDAV like all of the other calendar data--youe have
to use some other hacked up way to get it. Oh, and they deprecated
the WebDAV interface in favor of a SOAP interface. So eventually
we will have to support multiple protocols to be able to access the
wonder that is Microsoft Exchange Server calendaring.

I am negotiating access to a friend's MS Exchange account so that I
can try to get attendees working, but it may be a bit. I would just
re-write everything to use the SOAP interface, but it is only
available in Exchange Server 2007.  Gah.  But, it looks like
"Mainstream support" for ES 2003 ended 8 days ago...so maybe I should
rewrite it. To heck with those users of ES 2003. Although, apparently
"Extended support" lasts until 4/8/2014--whatever that means.

I really hope the new SOAP interface doesn't suck.