Outgoing PTP redirected calls did not wait for the COLR from the redirected-to party.
For outgoing PTP redirected calls, you now need to use the inhibit(i)
option on all of the REDIRECTING statements before dialing the redirected-to
party.  You still have to set the REDIRECTING(to-xxx,i) and the
REDIRECTING(from-xxx,i) values.  The PTP call will update the redirecting-to
presentation when it becomes available and queue the redirecting update to
the calling channel.