Sent wrong message to clear a call we started if the other end has not responed yet.
In the state MISDN_CALLING (i.e. SETUP was sent but no answer has arrived yet),
it is not allowed to clear the call with RELEASE_COMPLETE.  It must be
cleared with DISCONNECT.  A RELEASE_COMPLETE is only allowed as an answer
to a SETUP.  (See Q.931 ch. 5.3.2, 5.3.2.a, 5.3.2.b)

Patches:
    chan-misdn-ccstate7.patch uploaded by customer.

JIRA ABE-1862