Ensure that configure-script testing for compiler attributes actually works.
The configure script tests for compiler attributes didn't actually enable
enough warnings or provide a proper test harness to determine whether the 
compiler supports the attribute in question or not; this caused gcc 4.1 to
report that it supports 'weakref', but it doesn't actually support it in the
way that is needed for our optional API mechanism. The new configure script
test will properly distinguish between full support and partial support
for this attribute, among others.