Creating a branch to solve issue 12434.
Asterisk currently has a problem where it does not give proper answers
for all media streams offered in an SDP. For instance, in the issue 
mentioned above, the SDP offer contains a media stream for video. Since
video support is disabled, though, Asterisk provides no answer for the
video stream offered. Asterisk should have a corresponding m= line
in its answer that sets the port to 0.

For this branch, I will not be attempting to make any sweeping changes
to the SDP parser. I will simply be making sure that if we are going to
answer an SDP, that we have all media offers accounted for in the answer.